Fix Problem 6 Haskell
I realized that I accidentally had the terms flipped and got the negative
number. Since I was using this fancy ghc-mod thing in Atom that lets me
run the code inside the editor I didn't catch the difference. Thought it
was the prompt returning, not a negative sign. Now I know better.